Extraido codigo de tarjeta de noticia a componente reutilizable
