+++ src/components/searchLens/searchCategory/SearchCategory.jsx
@@ -1,8 +1,7 @@
 import { useEffect, useState } from 'react';
 import './SearchCategory.css';
-import Scorer from '../../Scorer';
 import useAuthHttpCall from '../../../hooks/useAuthHttpCall';
-import { BACKEND_URL } from '../../../config';
+import NewsCard from '../../NewsCard';
 
 const categories = [
   'Actualidad',
@@ -67,29 +66,7 @@ function SearchCategory() {
         <ul className="noticia">
           {news.map((noticia) => (
             <li className="newscard" key={noticia.id}>
-              <h2 className="newstitle">{noticia.title}</h2>
-              <h4 className="introtext">{noticia.introText} </h4>
-              <p className="text">{noticia.text}</p>
-              {noticia.imagenUrl && !noticia.imagenUrl.startsWith('http') ? (
-                <img
-                  className="imagen"
-                  src={`${BACKEND_URL}/uploads/${noticia.imagenUrl}`}
-                  alt={noticia.title}
-                />
-              ) : noticia.imagenUrl && noticia.imagenUrl.startsWith('http') ? (
-                <img
-                  className="imagen"
-                  src={noticia.imagenUrl}
-                  alt={noticia.title}
-                />
-              ) : null}
-              <div className="scorer">
-                <Scorer className="score" initial={noticia.score} />
-              </div>
-              <p className="date">
-                Fecha de publicación:{' '}
-                {new Date(noticia.publishDate).toLocaleDateString('es-ES')}
-              </p>
+              <NewsCard noticia={noticia} />
             </li>
           ))}
         </ul>
